improved setup script

+++ tasks/pfeed.rake
@@ -2,15 +2,23 @@ namespace :pfeed do
     
     desc 'Sets up parolkar\'s pfeed plugin '
     task :setup do
-      
+       root = "#{Rails.root}"
       raise 'pfeed plugin was only tested on unix & windows' if ! Rake.application.unix? &&  ! Rake.application.windows?
+      
+      if ! File.exists?("#{root}/vendor/plugins/inflectionist")
+        puts "Inflectionist plugin is required by pfeed, while you dont seem to have it installed \n Attempting to install..."
+        system "#{root}/script/plugin install git://github.com/parolkar/inflectionist.git "
+      end
+      
+      raise '...something went wrong please install http://github.com/parolkar/inflectionist first!' if ! File.exists?("#{root}/vendor/plugins/inflectionist")
+      
       
       files_to_be_copied = [
         {:source => "/vendor/plugins/pfeed/db/migrate/0000_create_pfeed_items.rb", :target => "/db/migrate/#{migration_timestamp}_create_pfeed_items.rb" },
          {:source => "/vendor/plugins/pfeed/db/migrate/0001_create_pfeed_deliveries.rb", :target => "/db/migrate/#{migration_timestamp}_create_pfeed_deliveries.rb" }
         ]
       
-      root = "#{Rails.root}"
+     
       FileUtils.mkdir_p("#{root}/db/migrate") unless File.exists?("#{root}/db/migrate")
       files_to_be_copied.each {|ftbc|
         FileUtils.cp_r(root+ftbc[:source], root+ftbc[:target]) #:force => true)
